Add bash script git-replacement.sh
By aliasing "git" to this script, we can add our own checks before
git does anything.

e.g. We can check if the user is trying to use
git push --force
and present them with an error message and explanation on why
it is a bad idea.

Once our checks are based, this script calls the real git program and
passes in the same parameters.

See #94