Add tab completion function for git lg (and gl)
This tab completion function is not necessary for git lg, since the
completion function for git log will be automatically used. However,
since our __git_shortcut() function in bash looks for the appropriate
autocomplete function when creating a bash alias, we need this function
to get the autocompletion for gl to work properly.

See #95