Add "git jira" command
The "git jira" command parses the branch name to determine the Jira
issue based on the naming scheme
{branchType}/{issue}-{description}
e.g. feature/sf-123-add-new-button (the issue number is "sf-123")

The git jira command uses this issue number in combination with the
custom git config value stored at 'salcode.jiraSubdomain' (in my example
this value is "ferrarello") to construct a
URL like

https://ferrarello.atlassian.net/browse/sf-123

and open this URL in the browser.

See #118